(PC-31328)[API] feat: invert decorators order for GET event stocks endpoint
